[ios]Hack: ページにより表示するデータを切り替えるロジックを共通化
