+++ client/Eventory/Codes/Controller/EventInfoViewController.swift
@@ -12,6 +12,10 @@ import SwiftTask
 
 class EventInfoViewController: BaseTableViewController {
 
+    override var viewPageClass: CheckStatus {
+        return CheckStatus.NoCheck
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
     }
@@ -20,29 +24,4 @@ class EventInfoViewController: BaseTableViewController {
         super.didReceiveMemoryWarning()
     }
 
-    override func refresh(completed: (() -> Void)? = nil) {
-        dispatch_async(dispatch_get_main_queue()) {
-            let task = [EventManager.sharedInstance.fetchNewEvent()]
-            Task.all(task).success { _ in
-                self.eventSummaries = EventManager.sharedInstance.getSelectNewEventAll()
-                completed?()
-                }.failure { _ in
-                    let alert: UIAlertController = UIAlertController(title: NetworkErrorTitle,message: NetworkErrorMessage, preferredStyle: .Alert)
-                    let cancelAction: UIAlertAction = UIAlertAction(title: NetworkErrorButton, style: .Cancel, handler: nil)
-                    alert.addAction(cancelAction)
-                    self.presentViewController(alert, animated: true, completion: nil)
-                    completed?()
-            }
-        }
-    }
-
-    override func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-        if let cell = self.tableView.dequeueReusableCellWithIdentifier(EventInfoTableViewCellIdentifier, forIndexPath: indexPath) as? EventInfoTableViewCell {
-            if let eventSummaries = self.eventSummaries {
-                cell.bind(eventSummaries[indexPath.row], viewPageClass: CheckStatus.Keep,indexPath: indexPath)
-                return cell
-            }
-        }
-        return UITableViewCell()
-    }
 }
